% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{addDataByDataFrame}
\alias{addDataByDataFrame}
\title{addDataByDataFrame}
\usage{
addDataByDataFrame(data = NULL, provenance = NULL, experiment = NULL,
  time_zone = "Europe/Paris")
}
\arguments{
\item{data}{data table}

\item{provenance}{provenance uri}

\item{time_zone}{provenance uri
"Variable","Date","Value" column are mandatory 
A column Sensor or ScientificObject is required
Both can be set

You can use as_tibble function
Data example: 
A tibble: 2 x 5
Date   ScientificObject       Variable          Value Sensor         
<chr>      <chr>                  <chr>             <dbl> <chr>          
2019-05-2… http://www.opensilex.… http://sinfonia.…  20   http://sinfoni…
2019-05-2… http://www.opensilex.… http://sinfonia.…  19.5 http://sinfoni…
Dataframe example with Scientific_object
\donttest{
connectToOpenSILEX( identifier="guest@opensilex.org",
              password="guest"
              url = "http://www.opensilex.org/openSilexAPI/rest")
  data = tibble::tribble(
   ~Date, ~ScientificObject,  ~Variable, ~Value,~Sensor,
  "2019-05-20 17:40:00", "http://www.opensilex.org/demo/DIA2017-1/so-test","http://www.opensilex.org/demo/set/variables#variable.t_t_t",  20.0, "http://www.opensilex.org/demo/set/devices/humidity-sensor-test",
  "2019-05-20 17:00:00", "http://www.opensilex.org/demo/DIA2017-1/so-test","http://www.opensilex.org/demo/set/variables#variable.t_t_t",  19.5, "http://www.opensilex.org/demo/set/devices/humidity-sensor-test"
  )
  addDataByDataFrame(
     data = data,
     provenance = "demo:id/provenance/1594993768653", 
     experiment= "demo:exp/sensor/s100000",
   time_zone = "Europe/Paris" 
  ) 
}}
}
\description{
Add  data from a tibble
}
