% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{addDataByExperiment}
\alias{addDataByExperiment}
\title{expData <- tibble::tibble(
    ScientificObject = c(
     "sunagri:SUA2011-1/so-test",
     "sunagri:SUA2011-1/so-test",
     "sunagri:SUA2011-1/so-test"),
    Date = c("2015-04-15", "2015-04-16", "2015-04-15T15:10:20+01:00"),
    Variable = c(
          "sunagri:id/variables/v148",
          "sunagri:id/variables/v148",
          "sunagri:id/variables/v148"
           ),
    Value = c(
     30.2,
     40.4,
     50.5
    )
)}
\usage{
addDataByExperiment(expData = NULL, experiment = NULL,
  provenance = NULL)
}
\description{
expData <- tibble::tibble(
    ScientificObject = c(
     "sunagri:SUA2011-1/so-test",
     "sunagri:SUA2011-1/so-test",
     "sunagri:SUA2011-1/so-test"),
    Date = c("2015-04-15", "2015-04-16", "2015-04-15T15:10:20+01:00"),
    Variable = c(
          "sunagri:id/variables/v148",
          "sunagri:id/variables/v148",
          "sunagri:id/variables/v148"
           ),
    Value = c(
     30.2,
     40.4,
     50.5
    )
)
}
\keyword{internal}
