% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{addSensorDataByCSV}
\alias{addSensorDataByCSV}
\title{addSensorDataByCSV}
\usage{
addSensorDataByCSV(file_path = NULL, delimiter = ",",
  start_column = NULL, provenance = NULL, experiment = NULL,
  time_zone = "Europe/Paris")
}
\arguments{
\item{file_path}{absolute csv file path}

\item{delimiter}{csv file delimiter}

\item{start_column}{csv start variable column number}

\item{provenance}{provenance uri}

\item{time_zone}{provenance uri
Datetime Column and variables column are mandatory

CSV example
,Air_temperature_20min_instant,Air_temperature_20min_instant
,Celsius_degree,Celsius_degree
,Canopy_air_temperature,Canopy_air_temperature
Sensor,http://www.opensilex.org/sunagri/2020/s20031,http://www.opensilex.org/sunagri/2020/s20031
DateTime,http://www.opensilex.org/sunagri/id/variables/v082,http://www.opensilex.org/sunagri/id/variables/v083
2019-05-20T15:40:00Z,19.971676,19.971676
2019-05-20T16:00:00Z,20.224501,20.224501
\donttest{
connectToOpenSILEX( identifier="guest@opensilex.org",
              password="guest"
              url = "http://www.opensilex.org/openSilexAPI/rest",)
  addSensorDataByCSV(
  start_column = 2,
  delimiter = ",",
  provenance = "demo:id/provenance/1594993768653",
   file_path = "/home/filpath/file.csv",
   time_zone = "Europe/Paris" 
  )
}
CSV example with Scientific object

;Method;Air_temperature_20min_instant
;Unit,Celsius_degree
;Entity + characteristic;Canopy_air_temperature
;Sensor;sunagri:2020/s20031
DateTime;Scientific_object;sunagri:id/variables/v151
2019-05-20T15:40:00Z;sunagri:2019/o19064783;19.971676
2019-05-20T16:00:00Z;sunagri:2019/o19064783;20.224501
2019-05-20T16:20:00Z;sunagri:2019/o19064783;19.466026

\donttest{
connectToOpenSILEX( identifier="guest@opensilex.org",
              password="guest"
              url = "http://www.opensilex.org/openSilexAPI/rest",)
  addSensorDataByCSV(
  start_column = 3,
  delimiter = ",",
  provenance = "demo:id/provenance/1594993768653",
   file_path = "/home/filpath/file.csv",
   time_zone = "Europe/Paris" 
  )
}}
}
\description{
Add sensor data from a formatted csv
}
