% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postAnnotations.R
\name{createAnnotation}
\alias{createAnnotation}
\title{postAnnotations}
\usage{
createAnnotation(newAnnotation)
}
\arguments{
\item{newAnnotation}{AnnotationCreationDTO}
}
\value{
WSResponse object
}
\description{
send a list of annotation to the web service
}
\details{
You have to execute the \code{\link{connectToOpenSILEXWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToOpenSILEXWS(
               url = "http://www.opensilex.org/rest/",
               username="guest@opensilex.org",
               password="guest")

newAnnotation <- AnnotationCreationDTO$new(
 targets = list("http://www.phenome-fppn.fr/test/2019/o19000074"),
 motivation  = "http://www.w3.org/ns/oa#describing",
 description =  "the object has been observed"
)

 response <- postAnnotations(newAnnotation)
 response$success
 response$metadata
 }
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any
         request on PHIS web service.
}
