% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{getExperimentDesign}
\alias{getExperimentDesign}
\title{getExperimentDesign}
\usage{
getExperimentDesign(uri = NULL)
}
\arguments{
\item{uri}{URI of the experiment}
}
\value{
WSResponse object
}
\description{
retrieves the informations for one experiment
}
\details{
You have to execute the \code{\link{connectToOpenSILEXWS}} function first to have access to the web
service
}
\examples{
\donttest{
 connectToOpenSILEX(
    identifier="guest@opensilex.org",
    password="guest", 
    url= "http://www.opensilex.org/openSilexAPI/rest") 
 Exp1<-getExperiment(
        experimentURI ="test-expe:test-serre")
 Exp1$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on OpenSILEX web service.
}
