% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiments.R
\name{getExperiments}
\alias{getExperiments}
\title{retrieves the experiments from the web service}
\usage{
getExperiments(name = "", year = "", is_ended = NULL,
  species = list(), factors = list(), projects = list(),
  is_public = NULL, order_by = "name=asc", page = NULL,
  page_size = NULL)
}
\arguments{
\item{name}{character, search by the name of an experiment (optional)}

\item{year}{character, search by the year of the experiment (optional)}

\item{is_ended}{boolean, search to the end date of experiment (optional)}

\item{species}{list, search experiment by list of species uri(s) (optional)}

\item{factors}{list, search experiment by list of factor uri(s) (optional)}

\item{projects}{list, search experiment by list of projects uri (optional)}

\item{is_public}{boolean, search experiment public status (optional)}

\item{order_by}{character, search by the place of an experiment (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{page_size}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
Retrieves the available experiments and/or linked to a project
}
\details{
You have to execute the \code{\link{connectToOpenSILEXWS}} function first to have access to the web
service
}
\examples{
\donttest{
 connectToOpenSILEXWS( 
                 username = "guestopensilex.org",
                  password = "guest",
   url = "http:/localhost:8666/rest")
 expes <- getExperiments()
 expes$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on OpenSILEX web service.
}
