% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{getVariablesByExperiment}
\alias{getVariablesByExperiment}
\title{getVariablesByExperiment}
\usage{
getVariablesByExperiment(uri = NULL)
}
\arguments{
\item{uri}{URI of the experiment}
}
\value{
WSResponse object
}
\description{
retrieves the informations variables used in one xperiment
}
\details{
You have to execute the \code{\link{connectToOpenSILEXWS}} function first to have access to the Web
Service
}
\examples{
\donttest{
 connectToOpenSILEXWS(username = "guest@opensilex.org",
                      password = "guest",
                      url = "http:/localhost:8666/rest")
 ExpVar<-getExperiment(experimentURI ="test-expe:test-serre")
 ExpVar$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on OpenSILEX web service.
}
